BR-14: refactor base structure
