+++ src/main/java/com/bobocode/svydovets/annotation/bean/processor/AutoSvydovetsBeanPostProcessor.java
@@ -1,29 +1,31 @@
 package com.bobocode.svydovets.annotation.bean.processor;
 
 import com.bobocode.svydovets.annotation.annotations.AutoSvydovets;
-import com.bobocode.svydovets.annotation.register.BeanRegistry;
-import com.bobocode.svydovets.annotation.register.DefaultBeanRegistry;
+import com.bobocode.svydovets.annotation.bean.factory.AnnotationBeanFactory;
+import com.bobocode.svydovets.annotation.bean.factory.BeanFactory;
+
 
 import java.lang.reflect.Field;
+import java.util.Map;
 
 //todo: provide JavaDocs
 //todo: perform refactoring
 public class AutoSvydovetsBeanPostProcessor implements BeanPostProcessor {
 
-    private final BeanRegistry beanRegistry;
+    private BeanFactory beanFactory;
 
     public AutoSvydovetsBeanPostProcessor() {
-        this.beanRegistry = new DefaultBeanRegistry();
+        this.beanFactory = new AnnotationBeanFactory();
     }
 
     @Override
-    public void process() {
-        for (var entrySet : beanRegistry.getRootContext().entrySet()) {
+    public void processBeans(Map<String, Object> rootContext) {
+        for (var entrySet : rootContext.entrySet()) {
             Object beanObject = entrySet.getValue();
             Class<?> beanType = beanObject.getClass();
             for (var field : beanType.getDeclaredFields()) {
                 if (field.isAnnotationPresent(AutoSvydovets.class)) {
-                    var dependency = beanRegistry.getSingleton(field.getType());
+                    var dependency = beanFactory.getBean(field.getType());
                     initField(beanObject, field, dependency);
                 }
             }
