--- src/main/java/com/bobocode/svydovets/support/AbstractApplicationContext.java
@@ -1,10 +1,0 @@
-package com.bobocode.svydovets.support;
-
-import com.bobocode.svydovets.annotation.bean.factory.BeanFactory;
-
-//todo: provide JavaDocs (optionally)
-public abstract class AbstractApplicationContext implements ApplicationContext {
-
-    protected abstract BeanFactory getBeanFactory();
-
-}
